# tests/golden/contracts/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Greenhouse API",
  "description": "Smart plant irrigation system with evidence-based plant care, multi-sensor conflict resolution, and self-learning irrigation profiles.\n\n",
  "version": "1.0.0"
 },
 "paths": {
...
  "/api/v1/plants/sync": {
   "post": {
    "tags": [
     "plants"
    ],
    "summary": "Sync Plants",
    "description": "Refresh plant care thresholds from the evidence-based plant database.\n\nResolves species to care data lookup and writes the result onto the matching\nplant rows. Use this after editing the bundled plant_database.json or after\nadding plants without explicit care fields. Scope is widening: single plant\nto cluster to entire database, depending on which field is set.\n\nArgs:\n    request: One of plant_id (single plant), cluster_id (all plants in\n        that cluster), or neither (every plant in every cluster).\n\nReturns:\n    Count of plants successfully synced and a list of error messages for\n    any that failed (failures do not abort the rest of the run).\n\nRaises:\n    HTTPException: 404 if plant_id is set and no such plant exists, or if\n        only cluster_id is set and no such cluster exists.",
    "operationId": "sync_plants",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SyncPlantsRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SyncPlantsResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
  "/api/v1/clusters/{cluster_id}/irrigate": {
   "post": {
    "tags": [
     "operations"
    ],
    "summary": "Irrigate",
    "description": "Run the full smart-irrigation pipeline for a cluster.\n\nSequence: sync sensors → resolve temperature (sensor / weather / fallback)\n→ run the decision engine → if not skipped and not a dry run, actuate the\ncluster's irrigator and record an `auto`-triggered event. Honors the\ncluster's irrigation config and the global 6h cooldown.\n\nSide effects: actuates physical hardware unless `dry_run=True`; writes a\nsensor sync and an irrigation event.\n\nArgs:\n    cluster_id: Cluster to irrigate.\n    request: `temp_override` to bypass sensor/weather resolution,\n        `dry_run=True` to compute the decision without actuating,\n        `no_sync=True` to skip the sensor refresh, `force=True` to\n        bypass the quiet-hours deny window (the decision still records\n        a `manual_override_quiet_hours` warning Reason).\n\nReturns:\n    Decision details (`action`, `reason`, `confidence`, optional\n    `duration_minutes`/`interval_hours`, plus the temperature used and its\n    source). `action` is one of `irrigated`, `skip`, or `error`.\n\nRaises:\n    HTTPException: 404 if the cluster does not exist.",
    "operationId": "irrigate",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/IrrigateRequest"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/IrrigateResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/clusters/{cluster_id}/monitor": {
   "get": {
    "tags": [
     "operations"
    ],
    "summary": "Monitor",
    "description": "Report each sensor's soil-moisture status and the plants that currently need water.\n\nUse this for sensor-only clusters (no irrigators) where you want to know\nwhich plants are dry without running the decision engine. Never actuates;\nsensors whose latest reading is stale are first refreshed from the Tuya\nCloud and the refreshed readings are stored.\n\nArgs:\n    cluster_id: Cluster to monitor.\n\nReturns:\n    Each sensor's current soil moisture, the plant's target band, and a\n    status (`ok` / `dry` / `very_dry` / `wet` / `no_data`), plus a flat\n    `needs_water` list of plant labels for the dry/very_dry sensors.\n\nRaises:\n    HTTPException: 404 if the cluster does not exist.",
    "operationId": "monitor",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MonitorResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/check": {
   "post": {
    "tags": [
     "operations"
    ],
    "summary": "Check All",
    "description": "Run a check across every cluster.\n\nFor each cluster: irrigate (if it has irrigators and `auto_run` is on) or\nmonitor (sensor-only). Collects learning + maintenance alerts per cluster.\n\nSide effects: may actuate physical hardware on any cluster whose decision\nengine says `irrigate` and whose config has `auto_run=True`. This is the\nsame call the background `check_all` job makes on its cron schedule\n(`IRRIGATION_CHECK_CRON_HOURS`, default hourly at :00).\n\nReturns:\n    Per-cluster results plus a `has_alerts` flag that is true if any\n    cluster has alerts, maintenance items, or thirsty plants.",
    "operationId": "check_all",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CheckAllResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/clusters/{cluster_id}/check": {
   "post": {
    "tags": [
     "operations"
    ],
    "summary": "Check Single",
    "description": "Run a check for a single cluster (irrigate or monitor + collect alerts).\n\nSide effects: may actuate physical hardware if the cluster has irrigators\nand `auto_run=True`. Same semantics as POST /check, scoped to one cluster.\n\nArgs:\n    cluster_id: Cluster to check.\n\nReturns:\n    The cluster's check result: irrigation or monitor outcome plus its\n    learning and maintenance alerts.\n\nRaises:\n    HTTPException: 404 if the cluster does not exist.",
    "operationId": "check_single",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "cluster_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Cluster Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/CheckClusterResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/sync": {
   "post": {
    "tags": [
     "operations"
    ],
    "summary": "Sync",
    "description": "Pull recent sensor readings from the Tuya Cloud into the local SQLite archive.\n\nThis is the same job the background scheduler runs every\n`IRRIGATION_SYNC_INTERVAL_MINUTES` (default 180); call\nthis manually after registering a new sensor or to backfill after an\noutage. No hardware is actuated.\n\nArgs:\n    request: `hours` of look-back to fetch (default 24).\n\nReturns:\n    Counts of total readings synced, new vs. duplicate-skipped, live\n    readings hit, and any per-sensor error messages.",
    "operationId": "sync",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/SyncRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/SyncResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "AlertResponse": {
    "properties": {
     "type": {
      "type": "string",
      "title": "Type"
     },
     "severity": {
      "type": "string",
      "title": "Severity"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "data": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Data"
     }
    },
    "type": "object",
    "required": [
     "type",
     "severity",
     "message"
    ],
    "title": "AlertResponse",
    "description": "An advisory alert raised alongside a decision, check or learning report."
   },
...
   "CheckAllResponse": {
    "properties": {
     "results": {
      "items": {
       "$ref": "#/components/schemas/CheckClusterResponse"
      },
      "type": "array",
      "title": "Results"
     },
     "has_alerts": {
      "type": "boolean",
      "title": "Has Alerts"
     }
    },
    "type": "object",
    "required": [
     "results",
     "has_alerts"
    ],
    "title": "CheckAllResponse",
    "description": "Outcome of the scheduled check across all clusters."
   },
   "CheckClusterResponse": {
    "properties": {
     "cluster_id": {
      "type": "integer",
      "title": "Cluster Id"
     },
     "cluster_name": {
      "type": "string",
      "title": "Cluster Name"
     },
     "action": {
      "type": "string",
      "title": "Action"
     },
     "notes": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Notes"
     },
     "alerts": {
      "items": {
       "$ref": "#/components/schemas/AlertResponse"
      },
      "type": "array",
      "title": "Alerts",
      "default": []
     },
     "maintenance": {
      "items": {
       "$ref": "#/components/schemas/AlertResponse"
      },
      "type": "array",
      "title": "Maintenance",
      "default": []
     },
     "needs_water": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Needs Water",
      "default": []
     }
    },
    "type": "object",
    "required": [
     "cluster_id",
     "cluster_name",
     "action"
    ],
    "title": "CheckClusterResponse",
    "description": "Outcome of the scheduled check for one cluster, with any alerts it raised."
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "IrrigateRequest": {
    "properties": {
     "temp_override": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temp Override"
     },
     "dry_run": {
      "type": "boolean",
      "title": "Dry Run",
      "default": false
     },
     "no_sync": {
      "type": "boolean",
      "title": "No Sync",
      "default": false
     },
     "force": {
      "type": "boolean",
      "title": "Force",
      "default": false
     }
    },
    "type": "object",
    "title": "IrrigateRequest",
    "description": "Body for POST /clusters/{id}/irrigate.\n\n``force=True`` is the explicit override switch used when the caller\nwants to bypass the quiet-hours deny window. The decision engine still\nruns every other rule (cooldown, stress, weather, sensors) and the\nfinal decision is tagged with a ``manual_override_quiet_hours`` warning\nReason so the audit log records the override."
   },
   "IrrigateResponse": {
    "properties": {
     "action": {
      "type": "string",
      "title": "Action"
     },
     "reason": {
      "type": "string",
      "title": "Reason"
     },
     "confidence": {
      "type": "number",
      "title": "Confidence"
     },
     "duration_minutes": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Duration Minutes"
     },
     "interval_hours": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Interval Hours"
     },
     "stress_indicators": {
      "anyOf": [
       {
        "additionalProperties": true,
        "type": "object"
       },
       {
        "type": "null"
       }
      ],
      "title": "Stress Indicators"
     },
     "reasons": {
      "items": {
       "$ref": "#/components/schemas/ReasonResponse"
      },
      "type": "array",
      "title": "Reasons",
      "default": []
     },
     "learning_alerts": {
      "items": {
       "$ref": "#/components/schemas/AlertResponse"
      },
      "type": "array",
      "title": "Learning Alerts",
      "default": []
     },
     "temperature": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temperature"
     },
     "temperature_source": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Temperature Source"
     }
    },
    "type": "object",
    "required": [
     "action",
     "reason",
     "confidence"
    ],
    "title": "IrrigateResponse",
    "description": "The decision for a cluster: action, duration, interval, confidence and the reason trail behind it."
   },
...
   "MonitorResponse": {
    "properties": {
     "cluster_name": {
      "type": "string",
      "title": "Cluster Name"
     },
     "sensors": {
      "items": {
       "$ref": "#/components/schemas/SensorStatusResponse"
      },
      "type": "array",
      "title": "Sensors"
     },
     "needs_water": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Needs Water"
     }
    },
    "type": "object",
    "required": [
     "cluster_name",
     "sensors",
     "needs_water"
    ],
    "title": "MonitorResponse",
    "description": "Moisture status of every sensor in a cluster and the plants that need water."
   },
...
   "ReasonResponse": {
    "properties": {
     "code": {
      "type": "string",
      "title": "Code"
     },
     "message": {
      "type": "string",
      "title": "Message"
     },
     "severity": {
      "type": "string",
      "title": "Severity",
      "default": "info"
     },
     "icon": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Icon"
     },
     "duration_delta": {
      "type": "integer",
      "title": "Duration Delta",
      "default": 0
     },
     "interval_delta": {
      "type": "integer",
      "title": "Interval Delta",
      "default": 0
     }
    },
    "type": "object",
    "required": [
     "code",
     "message"
    ],
    "title": "ReasonResponse",
    "description": "A single line in the decision's structured explanation trail."
   },
...
   "SensorStatusResponse": {
    "properties": {
     "sensor_id": {
      "type": "integer",
      "title": "Sensor Id"
     },
     "sensor_name": {
      "type": "string",
      "title": "Sensor Name"
     },
     "plant_species": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Species"
     },
     "soil_moisture": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Soil Moisture"
     },
     "status": {
      "type": "string",
      "title": "Status"
     },
     "target_min": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target Min"
     },
     "target_max": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Target Max"
     }
    },
    "type": "object",
    "required": [
     "sensor_id",
     "sensor_name",
     "plant_species",
     "soil_moisture",
     "status",
     "target_min",
     "target_max"
    ],
    "title": "SensorStatusResponse",
    "description": "One sensor's latest soil moisture judged against its plant's target band."
   },
...
   "SyncPlantsRequest": {
    "properties": {
     "plant_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Plant Id"
     },
     "cluster_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Cluster Id"
     }
    },
    "type": "object",
    "title": "SyncPlantsRequest",
    "description": "Scope of a plant-database sync: one plant, one cluster, or every plant when both are omitted."
   },
   "SyncPlantsResponse": {
    "properties": {
     "synced": {
      "type": "integer",
      "title": "Synced"
     },
     "errors": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Errors"
     }
    },
    "type": "object",
    "required": [
     "synced",
     "errors"
    ],
    "title": "SyncPlantsResponse",
    "description": "Number of plants refreshed from the plant database, plus per-plant errors."
   },
   "SyncRequest": {
    "properties": {
     "hours": {
      "type": "integer",
      "title": "Hours",
      "default": 24
     }
    },
    "type": "object",
    "title": "SyncRequest",
    "description": "Body for a sensor sync: how many hours of device history to backfill."
   },
   "SyncResponse": {
    "properties": {
     "total_synced": {
      "type": "integer",
      "title": "Total Synced"
     },
     "total_new": {
      "type": "integer",
      "title": "Total New"
     },
     "total_live": {
      "type": "integer",
      "title": "Total Live"
     },
     "errors": {
      "items": {
       "type": "string"
      },
      "type": "array",
      "title": "Errors"
     }
    },
    "type": "object",
    "required": [
     "total_synced",
     "total_new",
     "total_live",
     "errors"
    ],
    "title": "SyncResponse",
    "description": "Totals of a sensor sync (history rows synced, new rows stored, live reads) and any errors."
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 },
...
}